Move constraints out of model package
