+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/AgeRangeConstraintValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
