+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/DateNotInTheFuture.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -9,17 +9,13 @@ import javax.validation.Constraint;
 import javax.validation.Payload;
 
 @Documented
-@Constraint(validatedBy = AgeRangeConstraintValidator.class)
+@Constraint(validatedBy = DateNotInTheFutureConstraintValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AgeRangeValidator {
-    String message() default "Age must be between {minYears} and {maxYears}";
+public @interface DateNotInTheFuture {
+    String message() default "is in the future";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
-
-    int minYears() default 18;
-
-    int maxYears() default 150;
 }
