+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/DateNotInTheFutureConstraintValidator.java
@@ -1,13 +1,13 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import java.time.LocalDate;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
-public class DateNotInThePastConstraintValidator implements ConstraintValidator<DateNotInThePast, LocalDate> {
+public class DateNotInTheFutureConstraintValidator implements ConstraintValidator<DateNotInTheFuture, LocalDate> {
 
     @Override
-    public void initialize(DateNotInThePast dateNotInThePast) {
+    public void initialize(DateNotInTheFuture dateNotInTheFuture) {
         // NO-OP
     }
 
@@ -18,6 +18,6 @@ public class DateNotInThePastConstraintValidator implements ConstraintValidator<
             return true;
         }
 
-        return !localDate.isBefore(LocalDate.now());
+        return !localDate.isAfter(LocalDate.now());
     }
 }
