+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/DateNotInThePast.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -8,12 +8,12 @@ import java.lang.annotation.Target;
 import javax.validation.Constraint;
 import javax.validation.Payload;
 
-@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = {ValidCountyCourtJudgmentValidator.class})
 @Documented
-public @interface ValidCountyCourtJudgment {
-    String message() default "Invalid county court judgment request";
+@Constraint(validatedBy = DateNotInThePastConstraintValidator.class)
+@Target({ElementType.METHOD, ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface DateNotInThePast {
+    String message() default "is in the past";
 
     Class<?>[] groups() default {};
