+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/MobilePhoneNumber.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
