+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/constraints/MobilePhoneNumberConstraintValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
