+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/ClaimData.java
@@ -7,8 +7,8 @@ import cz.jirutka.validator.collection.constraints.EachNotNull;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.constraints.NotEmpty;
+import uk.gov.hmcts.cmc.domain.constraints.InterDependentFields;
 import uk.gov.hmcts.cmc.domain.models.amount.Amount;
-import uk.gov.hmcts.cmc.domain.models.constraints.InterDependentFields;
 import uk.gov.hmcts.cmc.domain.models.legalrep.StatementOfTruth;
 import uk.gov.hmcts.cmc.domain.models.otherparty.TheirDetails;
 import uk.gov.hmcts.cmc.domain.models.particulars.HousingDisrepair;
