+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/Interest.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.cmc.domain.models;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
-import uk.gov.hmcts.cmc.domain.models.constraints.InterDependentFields;
+import uk.gov.hmcts.cmc.domain.constraints.InterDependentFields;
 
 import java.math.BigDecimal;
 import java.util.Objects;
