+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/InterestDate.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import uk.gov.hmcts.cmc.domain.models.constraints.DateNotInTheFuture;
-import uk.gov.hmcts.cmc.domain.models.constraints.InterDependentFields;
+import uk.gov.hmcts.cmc.domain.constraints.DateNotInTheFuture;
+import uk.gov.hmcts.cmc.domain.constraints.InterDependentFields;
 
 import java.time.LocalDate;
 import java.util.Objects;
