+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/amount/AmountRange.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.cmc.domain.models.amount;
 
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
-import uk.gov.hmcts.cmc.domain.models.constraints.Money;
+import uk.gov.hmcts.cmc.domain.constraints.Money;
 
 import java.math.BigDecimal;
 import java.util.Objects;
