+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/party/Individual.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.cmc.domain.models.party;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
+import uk.gov.hmcts.cmc.domain.constraints.AgeRangeValidator;
 import uk.gov.hmcts.cmc.domain.models.Address;
-import uk.gov.hmcts.cmc.domain.models.constraints.AgeRangeValidator;
 import uk.gov.hmcts.cmc.domain.models.legalrep.Representative;
 
 import java.time.LocalDate;
