+++ domain-model/src/test/java/uk/gov/hmcts/cmc/domain/constraints/DateNotInThePastValidatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models.constraints;
+package uk.gov.hmcts.cmc.domain.constraints;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
